refactor: isolate resolvable string in a different namespace
