+++ resources/js/components/common/EnvironmentAwareInput.vue
@@ -8,7 +8,7 @@ import { AppInput, inputVariants, type InputVariants } from '@/components/base/i
 import { EnvVariableSegment } from '@/components/common/EnvVariableSegment';
 import { useEnvVariablesAwareString } from '@/composables/request/useEnvVariablesAwareString';
 import { useTabHorizontalScroll } from '@/composables/ui/useTabHorizontalScroll';
-import type { ResolvableString } from '@/interfaces/http';
+import type { ResolvableString } from '@/interfaces/common/resolvable';
 import { EnvVariableCheckStatus } from '@/utils/request/environment-variable-resolver';
 import { cn } from '@/utils/ui';
 import { useForwardExpose, useForwardPropsEmits } from 'reka-ui';
