+++ resources/js/components/domain/Client/Request/RequestAuthorization/RequestAuthorizationBasicAuth.vue
@@ -4,7 +4,7 @@
  * @description Input fields for Basic Authentication (username/password).
  */
 import EnvironmentAwareInput from '@/components/common/EnvironmentAwareInput.vue';
-import type { ResolvableString } from '@/interfaces/http';
+import type { ResolvableString } from '@/interfaces/common/resolvable';
 
 /*
  * Types & Interfaces.
