+++ resources/js/components/domain/Client/Request/RequestBody/RequestBodyContent.vue
@@ -4,7 +4,8 @@
  * @description Dynamic content renderer for the request body based on the selected payload type.
  */
 import { AppScrollArea } from '@/components/base/scroll-area';
-import { RequestBodyTypeEnum, type ResolvableString } from '@/interfaces/http';
+import { RequestBodyTypeEnum } from '@/interfaces/http';
+import type { ResolvableString } from '@/interfaces/common/resolvable';
 import type { JSONSchema7 } from 'json-schema';
 import RequestBodyFormData from './RequestBodyFormData.vue';
 import RequestBodyFormNone from './RequestBodyFormNone.vue';
