+++ resources/js/interfaces/common/resolvable.ts
@@ -1,0 +1,5 @@
+/**
+ * Represents a string that can exist either as a raw value (with potential placeholders)
+ * or as its final resolved version after environment variables are processed.
+ */
+export type ResolvableString = string | { raw: string; resolved: string };
