+++ resources/js/tests/utils/request/environment-variable-resolver.test.ts
@@ -7,7 +7,7 @@ import {
     getStringSegments,
     replaceEnvVariablesInString,
 } from '@/utils/request/environment-variable-resolver';
-import { resolveResolvableString } from '@/utils/request/resolvable-value';
+import { resolveResolvableString } from '@/utils/common/resolvable';
 import { describe, expect, it } from 'vitest';
 
 const createEnvVariablesMap = (variables: ParameterContract[]) => {
